scripts/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-geohash==0.8.5
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0





